requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements.in
#
aiodataloader==0.2.0
    # via -r requirements.in
amqp==2.6.1
    # via kombu
analytics-python==1.3.0b1
    # via shared
appdirs==1.4.4
    # via virtualenv
ariadne-django==0.2.0
    # via -r requirements.in
ariadne==0.13.0
    # via
    #   -r requirements.in
    #   ariadne-django
asgiref==3.3.4
    # via
    #   django
    #   opentelemetry-util-http
astroid==2.6.6
    # via pylint
attrs==20.3.0
    # via pytest
backoff==1.6.0
    # via analytics-python
billiard==3.6.3.0
    # via celery
boto3==1.9.218
    # via shared
botocore==1.12.253
    # via
    #   boto3
    #   s3transfer
cachetools==4.1.1
    # via google-auth
celery==4.3.0
    # via -r requirements.in
cerberus==1.3.2
    # via
    #   -r requirements.in
    #   shared
certifi==2020.6.20
    # via
    #   elastic-apm
    #   httpx
    #   minio
    #   requests
    #   sentry-sdk
cffi==1.14.2
    # via
    #   cryptography
    #   google-crc32c
cfgv==3.2.0
    # via pre-commit
chardet==3.0.4
    # via requests
git+ssh://git@github.com/codecov/opentelem-python.git@v0.0.4a1#egg=codecovopentelem
    # via -r requirements.in
colour==0.1.5
    # via shared
configparser==5.0.0
    # via minio
coverage==6.0.1
    # via
    #   codecovopentelem
    #   pytest-cov
cryptography==35.0.0
    # via shared
ddtrace==0.46
    # via -r requirements.in
deprecated==1.2.12
    # via opentelemetry-api
distlib==0.3.1
    # via virtualenv
django-cors-headers==3.7.0
    # via -r requirements.in
https://github.com/photocrowd/django-cursor-pagination/archive/f560902696b0c8509e4d95c10ba0d62700181d84.tar.gz
    # via -r requirements.in
django-dynamic-fixture==3.1.1
    # via -r requirements.in
django-filter==2.4.0
    # via -r requirements.in
django==3.1.13
    # via
    #   -r requirements.in
    #   ariadne-django
    #   django-cors-headers
    #   django-filter
    #   djangorestframework
djangorestframework==3.12.2
    # via -r requirements.in
docutils==0.15.2
    # via botocore
ecdsa==0.16.0
    # via tlslite-ng
elastic-apm==6.3.3
    # via -r requirements.in
environs==7.2.0
    # via -r requirements.in
factory-boy==3.2.0
    # via -r requirements.in
faker==4.1.3
    # via factory-boy
fakeredis==1.5.0
    # via -r requirements.in
filelock==3.0.12
    # via virtualenv
freezegun==1.1.0
    # via -r requirements.in
google-api-core==1.22.2
    # via google-cloud-core
google-auth==1.21.2
    # via
    #   google-api-core
    #   google-cloud-storage
google-cloud-core==1.4.1
    # via google-cloud-storage
google-cloud-storage==1.38.0
    # via shared
google-crc32c==1.0.0
    # via google-resumable-media
google-resumable-media==1.2.0
    # via google-cloud-storage
googleapis-common-protos==1.52.0
    # via google-api-core
graphql-core==3.1.3
    # via ariadne
gunicorn==20.0.4
    # via -r requirements.in
h11==0.12.0
    # via httpcore
httpcore==0.12.3
    # via httpx
httplib2==0.18.1
    # via oauth2
httpx==0.16.1
    # via shared
identify==2.2.2
    # via pre-commit
idna==2.8
    # via
    #   requests
    #   rfc3986
    #   yarl
iniconfig==1.1.1
    # via pytest
intervaltree==3.1.0
    # via ddtrace
isort==5.9.3
    # via pylint
jmespath==0.10.0
    # via
    #   boto3
    #   botocore
kombu==4.6.11
    # via celery
lazy-object-proxy==1.6.0
    # via astroid
marshmallow==3.8.0
    # via environs
mccabe==0.6.1
    # via pylint
minio==6.0.0
    # via
    #   -r requirements.in
    #   shared
monotonic==1.5
    # via analytics-python
multidict==4.7.6
    # via yarl
nodeenv==1.5.0
    # via pre-commit
oauth2==1.9.0.post1
    # via shared
oauthlib==3.1.0
    # via shared
opentelemetry-api==1.3.0
    # via
    #   opentelemetry-instrumentation
    #   opentelemetry-instrumentation-django
    #   opentelemetry-instrumentation-wsgi
    #   opentelemetry-sdk
    #   opentelemetry-util-http
opentelemetry-instrumentation-django==0.22b0
    # via -r requirements.in
opentelemetry-instrumentation-wsgi==0.22b0
    # via opentelemetry-instrumentation-django
opentelemetry-instrumentation==0.22b0
    # via
    #   opentelemetry-instrumentation-django
    #   opentelemetry-instrumentation-wsgi
    #   opentelemetry-util-http
opentelemetry-sdk==1.3.0
    # via
    #   -r requirements.in
    #   codecovopentelem
opentelemetry-semantic-conventions==0.22b0
    # via
    #   opentelemetry-instrumentation-django
    #   opentelemetry-instrumentation-wsgi
    #   opentelemetry-sdk
opentelemetry-util-http==0.22b0
    # via opentelemetry-instrumentation-django
opentracing==2.4.0
    # via -r requirements.in
packaging==20.9
    # via pytest
pluggy==0.13.1
    # via pytest
pre-commit==2.11.1
    # via -r requirements.in
protobuf==3.13.0
    # via
    #   ddtrace
    #   google-api-core
    #   googleapis-common-protos
psycopg2==2.8.6
    # via -r requirements.in
py==1.9.0
    # via pytest
pyasn1-modules==0.2.8
    # via google-auth
pyasn1==0.4.8
    # via
    #   pyasn1-modules
    #   rsa
pycparser==2.20
    # via cffi
pylint-django==2.4.4
    # via -r requirements.in
pylint-plugin-utils==0.6
    # via pylint-django
pylint==2.9.6
    # via
    #   -r requirements.in
    #   pylint-django
    #   pylint-plugin-utils
pyparsing==2.4.7
    # via packaging
pytest-cov==2.12.1
    # via -r requirements.in
pytest-django==4.4.0
    # via -r requirements.in
pytest-mock==1.12.1
    # via -r requirements.in
pytest==6.2.5
    # via
    #   -r requirements.in
    #   pytest-cov
    #   pytest-django
    #   pytest-mock
python-dateutil==2.7.5
    # via
    #   -r requirements.in
    #   analytics-python
    #   botocore
    #   faker
    #   freezegun
    #   minio
python-dotenv==0.14.0
    # via environs
python-json-logger==0.1.11
    # via -r requirements.in
python-redis-lock==3.7.0
    # via -r requirements.in
pytz==2018.7
    # via
    #   -r requirements.in
    #   celery
    #   django
    #   google-api-core
    #   minio
pyyaml==5.1
    # via
    #   pre-commit
    #   vcrpy
redis==3.2.0
    # via
    #   -r requirements.in
    #   fakeredis
    #   python-redis-lock
requests==2.25.1
    # via
    #   -r requirements.in
    #   analytics-python
    #   codecovopentelem
    #   google-api-core
    #   google-cloud-storage
    #   stripe
rfc3986[idna2008]==1.4.0
    # via httpx
rsa==4.7.2
    # via google-auth
s3transfer==0.2.1
    # via boto3
sentry-sdk==1.4.3
    # via -r requirements.in
setproctitle==1.1.10
    # via -r requirements.in
git+ssh://git@github.com/codecov/shared.git@v0.7.12#egg=shared
    # via -r requirements.in
simplejson==3.17.2
    # via -r requirements.in
six==1.15.0
    # via
    #   analytics-python
    #   cryptography
    #   django-dynamic-fixture
    #   ecdsa
    #   fakeredis
    #   google-api-core
    #   google-auth
    #   google-resumable-media
    #   protobuf
    #   python-dateutil
    #   tenacity
    #   vcrpy
    #   virtualenv
sniffio==1.2.0
    # via
    #   httpcore
    #   httpx
sortedcontainers==2.2.2
    # via
    #   fakeredis
    #   intervaltree
sqlparse==0.4.1
    # via django
starlette==0.14.2
    # via ariadne
statsd==3.3.0
    # via shared
stripe==2.55.2
    # via -r requirements.in
tenacity==6.2.0
    # via ddtrace
text-unidecode==1.3
    # via faker
tlslite-ng==0.7.5
    # via
    #   -r requirements.in
    #   shared
toml==0.10.2
    # via
    #   pre-commit
    #   pylint
    #   pytest
    #   pytest-cov
typing-extensions==3.7.4.3
    # via ariadne
urllib3==1.25.10
    # via
    #   botocore
    #   elastic-apm
    #   minio
    #   requests
    #   sentry-sdk
vcrpy==2.0.1
    # via -r requirements.in
vine==1.3.0
    # via
    #   amqp
    #   celery
virtualenv==20.4.3
    # via pre-commit
whitenoise==5.2.0
    # via -r requirements.in
wrapt==1.12.1
    # via
    #   astroid
    #   deprecated
    #   opentelemetry-instrumentation
    #   vcrpy
yarl==1.5.1
    # via vcrpy

# The following packages are considered to be unsafe in a requirements file:
# setuptools
